split brightcove
